third_party/WebKit/LayoutTests/http/tests/devtools/search/source-frame-search-expected.txt

Tests different types of search in SourceFrame


Running: testSearch
Running search test for query = searchTestUniqueString:
  - function <searchTestUniqueString>()
  -     // <searchTestUniqueString> two occurences on the same line searchTestUniqueString
  -     // searchTestUniqueString two occurences on the same line <searchTestUniqueString>
  -     // <searchTestUniqueString> on the next line.
  -     <searchTestUniqueString>();
  -     // <SEARCHTestUniqueString>();

Running: testSearchCaseSensitive
Running search test for query = SEARCHTestUniqueString (caseSensitive):
  -     // <SEARCHTestUniqueString>();

Running: testSearchRegex
Running search test for query = searchTestUnique.* (regex):
  - function <searchTestUniqueString()>
  -     // <searchTestUniqueString two occurences on the same line searchTestUniqueString>
  -     // <searchTestUniqueString on the next line.>
  -     <searchTestUniqueString();>
  -     // <SEARCHTestUniqueString();>
  - // <searchTestUnique space String>

Running: testSearchCaseSensitiveRegex
Running search test for query = searchTestUnique.* (regex, caseSensitive):
  - function <searchTestUniqueString()>
  -     // <searchTestUniqueString two occurences on the same line searchTestUniqueString>
  -     // <searchTestUniqueString on the next line.>
  -     <searchTestUniqueString();>
  - // <searchTestUnique space String>

Running: testSearchConsequent
Running search test for query = AAAAA:
  - // <AAAAA>AAAAAA 11xA here
  - // AAAAA<AAAAA>A 11xA here
